Rag_app/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.8
pymysql>=1.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
plotly>=5.0.0
pydantic>=2.0.0
bcrypt>=4.0.0
urllib3>=1.26.0
google-generativeai>=0.3.0